requirements.txt
tifffile==2022.8.12
scikit-image==0.19.3
PyWavelets==1.3.0
pystackreg==0.2.6.post1
PyOpenGl==3.1.6
pyimgui-wheels==2.0.1
Pillow==9.2.0
pandas==1.4.4
numpy==1.23.2
matplotlib==3.5.3
joblib==1.1.0
glfw==2.5.5
colorcet==3.0.0
dill==0.3.5.1
psutil==5.9.2
opencv-python==4.6.0.66
pyperclip==1.8.2